add comfyui_instantid/README.md
