+++ onediff_comfy_nodes/modules/oneflow/__init__.py
@@ -2,3 +2,4 @@ from .config import _USE_UNET_INT8, ONEDIFF_QUANTIZED_OPTIMIZED_MODELS
 from .booster_basic import BasicOneFlowBoosterExecutor
 from .booster_deepcache import DeepcacheBoosterExecutor
 from .booster_patch import PatchBoosterExecutor
+from .patch_for_oneflow import *
