+++ onediff_comfy_nodes/modules/oneflow/hijack_ipadapter_plus/IPAdapterPlus.py
@@ -1,14 +1,14 @@
 """hijack ComfyUI/custom_nodes/ComfyUI_IPAdapter_plus/IPAdapterPlus.py"""
 import os
-from pathlib import Path
+
+from register_comfy.CrossAttentionPatch import \
+    CrossAttentionPatch as CrossAttentionPatch_OF
 
 from onediff.infer_compiler.deployable_module import DeployableModule
 from onediff.infer_compiler.transform import torch2oflow
 
-from ._config import (ipadapter_plus_hijacker, ipadapter_plus_of,
-                      ipadapter_plus_pt)
-from .CrossAttentionPatch import CrossAttentionPatch as CrossAttentionPatch_OF
 from ..utils.booster_utils import is_using_oneflow_backend
+from ._config import ipadapter_plus_hijacker, ipadapter_plus_pt
 
 set_model_patch_replace_fn_pt = ipadapter_plus_pt.IPAdapterPlus.set_model_patch_replace
