+++ onediff_comfy_nodes/modules/oneflow/hijack_ipadapter_plus/__init__.py
@@ -2,4 +2,4 @@ from ._config import ipadapter_plus_hijacker, is_load_ipadapter_plus_pkg
 
 if is_load_ipadapter_plus_pkg:
     from .IPAdapterPlus import *
-    from .patch_for_oneflow import *
+    from ..patch_for_oneflow import *
